Use double for equilibrium inputs.

+++ linearization/@linss/linearize.m
@@ -115,7 +115,7 @@ B = matlabFunction(symfun(dfdu, [X; U; Mu]));
 C = matlabFunction(symfun(dhdx, [X; U; Mu]));
 D = matlabFunction(symfun(dhdu, [X; U; Mu]));
 
-obj = linss(A, B, C, D, x0, u0, mu0);
+obj = linss(A, B, C, D, double(x0), double(u0), double(mu0));
 
 end
